// YOUCOM.ReserVook/src/app/features/web/master/scname/model/scname.model.ts
export class ScNameInfo {
  companyNo: string;
  scCd: string;
  scCategoryCd: string;
  scSegCd: string;
  content1: string;
  content2: string;
  displayOdr: number;
  updateCnt: number;
  programId: string;
  createClerk: string;
  createMachineNo: string;
  createMachine: string;
  createDatetime: string;
  updateClerk: string;
  updateMachineNo: string;
  updateMachine: string;
  updateDatetime: string;
  status: string;
}